dbeaver/pro#7023 refactor: data grid search state management and persistence
