+++ webapp/packages/plugin-data-spreadsheet-new/src/module.ts
@@ -1,6 +1,6 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2025 DBeaver Corp and others
+ * Copyright (C) 2020-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ import { DataGridContextMenuSaveContentService } from './DataGrid/DataGridContex
 import { DataGridContextMenuOrderService } from './DataGrid/DataGridContextMenu/DataGridContextMenuOrderService.js';
 import { DataGridContextMenuFilterService } from './DataGrid/DataGridContextMenu/DataGridContextMenuFilter/DataGridContextMenuFilterService.js';
 import { DataGridContextMenuCellEditingService } from './DataGrid/DataGridContextMenu/DataGridContextMenuCellEditingService.js';
-import { SearchPersistenceService } from './SearchPersistenceService.js';
+import { SearchStateService } from './DataGridSearchStateService.js';
 
 export default ModuleRegistry.add({
   name: '@cloudbeaver/plugin-data-spreadsheet-new',
@@ -25,7 +25,7 @@ export default ModuleRegistry.add({
       .addSingleton(Bootstrap, SpreadsheetBootstrap)
       .addSingleton(Bootstrap, LocaleService)
       .addSingleton(DataGridContextMenuCellEditingService)
-      .addSingleton(SearchPersistenceService)
+      .addSingleton(SearchStateService)
       .addSingleton(DataGridSettingsService)
       .addSingleton(DataGridContextMenuSaveContentService)
       .addSingleton(DataGridContextMenuOrderService)
